Change file structure.
